requirements.txt
apify-client==1.12.0
pandas==2.2.2
python-dotenv==1.0.1
google-cloud-translate==3.12.0
openai==1.93.0
anthropic==0.28.0
tiktoken 